intermediate commit before more refactoring

+++ src/include/hadron/Type.hpp
@@ -8,7 +8,7 @@ namespace hadron {
 // These are deliberately independent bits to allow for quick aggregate type comparisons, such as
 // type & (kInteger | kFloat) to determine if a type is numeric or
 // type & (kString | kSymbol) for character types, etc.
-enum Type : std::uint_fast16_t {
+enum Type : std::int32_t {
     kNil     = 0x0001,
     kInteger = 0x0002,
     kFloat   = 0x0004,
